+++ pkg/auth/handler/webapp/authflowv2/deps.go
@@ -40,4 +40,5 @@ var DependencySet = wire.NewSet(
 	wire.Struct(new(AuthflowV2WechatHandler), "*"),
 	wire.Struct(new(AuthflowV2AccountLinkingHandler), "*"),
 	wire.Struct(new(AuthflowV2LDAPLoginHandler), "*"),
+	wire.Struct(new(SettingsV2Handler), "*"),
 )
