--- pkg/auth/handler/webapp/settingsv2/deps.go
@@ -1,9 +1,0 @@
-package settingsv2
-
-import (
-	"github.com/google/wire"
-)
-
-var DependencySet = wire.NewSet(
-	wire.Struct(new(SettingsV2Handler), "*"),
-)
