--- pkg/auth/handler/webapp/settingsv2/settings.go
@@ -1,78 +1,0 @@
-package settingsv2
-
-import (
-	"net/http"
-
-	handlerwebapp "github.com/authgear/authgear-server/pkg/auth/handler/webapp"
-	"github.com/authgear/authgear-server/pkg/auth/handler/webapp/viewmodels"
-	"github.com/authgear/authgear-server/pkg/lib/session"
-	"github.com/authgear/authgear-server/pkg/util/httproute"
-	"github.com/authgear/authgear-server/pkg/util/template"
-)
-
-var TemplateWebSettingsV2HTML = template.RegisterHTML(
-	"web/settingsv2/settings.html",
-	handlerwebapp.Components...,
-)
-
-func ConfigureSettingsV2Route(route httproute.Route) httproute.Route {
-	return route.
-		WithMethods("OPTIONS", "GET", "POST").
-		WithPathPattern("/settings")
-}
-
-type SettingsV2Handler struct {
-	ControllerFactory        handlerwebapp.ControllerFactory
-	BaseViewModel            *viewmodels.BaseViewModeler
-	AuthenticationViewModel  *viewmodels.AuthenticationViewModeler
-	SettingsProfileViewModel *viewmodels.SettingsProfileViewModeler
-	Identities               handlerwebapp.SettingsIdentityService
-	Renderer                 handlerwebapp.Renderer
-}
-
-func (h *SettingsV2Handler) GetData(r *http.Request, rw http.ResponseWriter) (map[string]interface{}, error) {
-	userID := session.GetUserID(r.Context())
-
-	data := map[string]interface{}{}
-
-	// BaseViewModel
-	baseViewModel := h.BaseViewModel.ViewModel(r, rw)
-	viewmodels.Embed(data, baseViewModel)
-
-	// SettingsProfileViewModel
-	profileViewModelPtr, err := h.SettingsProfileViewModel.ViewModel(*userID)
-	if err != nil {
-		return nil, err
-	}
-	viewmodels.Embed(data, *profileViewModelPtr)
-
-	// Identity - Part 1
-	candidates, err := h.Identities.ListCandidates(*userID)
-	if err != nil {
-		return nil, err
-	}
-	authenticationViewModel := h.AuthenticationViewModel.NewWithCandidates(candidates, r.Form)
-	viewmodels.Embed(data, authenticationViewModel)
-
-	return data, nil
-}
-
-func (h *SettingsV2Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	ctrl, err := h.ControllerFactory.New(r, w)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-	defer ctrl.Serve()
-
-	ctrl.Get(func() error {
-		data, err := h.GetData(r, w)
-		if err != nil {
-			return err
-		}
-
-		h.Renderer.RenderHTML(w, r, TemplateWebSettingsV2HTML, data)
-
-		return nil
-	})
-}
