+++ pkg/lib/web/html.go
@@ -4,6 +4,13 @@ import (
 	"github.com/authgear/authgear-server/pkg/util/template"
 )
 
+func init() {
+	ComponentsHTML = append(ComponentsHTML, BaseComponentsHTML...)
+	ComponentsHTML = append(ComponentsHTML, TemplateWebAuthflowV2PageFrameHTML)
+	SettingsComponentsHTML = append(ComponentsHTML, BaseComponentsHTML...)
+	SettingsComponentsHTML = append(ComponentsHTML, TemplateWebSettingsV2PageFrameHTML)
+}
+
 var TemplateWebHTMLHeadHTML = template.RegisterHTML("web/__html_head.html")
 var TemplateWebGeneratedAssetHTML = template.RegisterHTML("web/__generated_asset.html")
 var TemplateWebHeaderHTML = template.RegisterHTML("web/__header.html")
@@ -59,16 +66,11 @@ var TemplateWebAuthflowV2ErrorPageLayoutHTML = template.RegisterHTML("web/authfl
 var TemplateWebAuthflowV2DeviceTokenCheckboxHTML = template.RegisterHTML("web/authflowv2/__device_token_checkbox.html")
 var TemplateWebAuthflowV2TermsOfServiceAndPrivacyPolicyFooterHTML = template.RegisterHTML("web/authflowv2/__toc_pp_footer.html")
 var TemplateWebAuthflowV2WatermarkHTML = template.RegisterHTML("web/authflowv2/__watermark.html")
-
-var TemplateWebSettingsV2LayoutHTML = template.RegisterHTML("web/settingsv2/layout.html")
-
-var TemplateWebSettingsV2WidePageFrameHTML = template.RegisterHTML("web/settingsv2/__wide_page_frame.html")
-var TemplateWebSettingsV2HTMLHeadHTML = template.RegisterHTML("web/settingsv2/__html_head.html")
-var TemplateWebSettingsV2GeneratedAssetHTML = template.RegisterHTML("web/settingsv2/__generated_asset.html")
-var TemplateWebSettingsV2LockoutHTML = template.RegisterHTML("web/settingsv2/__lockout.html")
 var TemplateWebAuthflowV2CSRFErrorPageLayoutHTML = template.RegisterHTML("web/authflowv2/__csrf_error_page_layout.html")
 
-var ComponentsHTML = []*template.HTML{
+var TemplateWebSettingsV2PageFrameHTML = template.RegisterHTML("web/authflowv2/__settings_page_frame.html")
+
+var BaseComponentsHTML = []*template.HTML{
 	TemplateWebHTMLHeadHTML,
 	TemplateWebGeneratedAssetHTML,
 	TemplateWebHeaderHTML,
@@ -97,7 +99,6 @@ var ComponentsHTML = []*template.HTML{
 	TemplateWebAuthflowV2LoadBotProtectionHTML,
 	TemplateWebAuthflowV2GeneratedAssetHTML,
 	TemplateWebAuthflowV2BasePageFrameHTML,
-	TemplateWebAuthflowV2PageFrameHTML,
 	TemplateWebAuthflowV2DialogHTML,
 	TemplateWebAuthflowV2BotProtectionWidgetHTML,
 	TemplateWebAuthflowV2BotProtectionFormInputHTML,
@@ -122,11 +123,8 @@ var ComponentsHTML = []*template.HTML{
 	TemplateWebAuthflowV2DeviceTokenCheckboxHTML,
 	TemplateWebAuthflowV2TermsOfServiceAndPrivacyPolicyFooterHTML,
 	TemplateWebAuthflowV2WatermarkHTML,
-
-	TemplateWebSettingsV2LayoutHTML,
-	TemplateWebSettingsV2WidePageFrameHTML,
-	TemplateWebSettingsV2HTMLHeadHTML,
-	TemplateWebSettingsV2GeneratedAssetHTML,
-	TemplateWebSettingsV2LockoutHTML,
 	TemplateWebAuthflowV2CSRFErrorPageLayoutHTML,
 }
+
+var ComponentsHTML []*template.HTML
+var SettingsComponentsHTML []*template.HTML
