--- resources/authgear/templates/en/web/settingsv2/layout.html
@@ -1,5 +1,0 @@
-<div class="layout--default">
-  <div class="layout__header--default"></div>
-  {{ template "wide_widget" . }}
-  <div class="layout__footer--default"></div>
-</div>
